feat(orga): make sitemaps with jsx
Using https://yarnpkg.com/en/package/jsx-string to make sitemaps creates a way cleaner code! This requires to be set up with babel for the jsx pragma.

A change internally as well is that sitemapIndex is not mutating the internal representation anymore, but rather left in index.js to push into an array that is eventually used to render the sitemap-index.xml